Refactored services files
