+++ learning_projects/MovieAppCleanArchitecture/app/src/main/java/com/example/movieappcleanarchitecture/data/remote/api/impl/BaseApiService.kt
@@ -1,12 +1,9 @@
-package com.example.movieappcleanarchitecture.services
+package com.example.movieappcleanarchitecture.data.remote.api.impl
 
 import com.example.movieappcleanarchitecture.common.API_BASE_URL
-import com.example.movieappcleanarchitecture.data.remote.api.MovieApiService
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-private val retrofit =
+val retrofit: Retrofit =
     Retrofit.Builder().baseUrl(API_BASE_URL).addConverterFactory(GsonConverterFactory.create())
         .build()
-
-val movieService: MovieApiService = retrofit.create(MovieApiService::class.java)
