+++ learning_projects/MovieAppCleanArchitecture/app/src/main/java/com/example/movieappcleanarchitecture/data/remote/api/impl/MovieApiServiceImpl.kt
@@ -1,0 +1,5 @@
+package com.example.movieappcleanarchitecture.data.remote.api.impl
+
+import com.example.movieappcleanarchitecture.data.remote.api.MovieApiService
+
+val movieService: MovieApiService = retrofit.create(MovieApiService::class.java)
